server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
google-genai==0.2.2
pypdf2==3.0.1
python-multipart==0.0.12

